[fix] : mu and sigma ok big changes coming to mode and maybe operator too, so commit before

+++ inst/include/snipsnop/SNPvectorDisk.h
@@ -44,7 +44,7 @@ class SNPVectorDisk : public SNPvector {
     return mode_;
   }
   
-  uint8_t mode(unsigned int n) {
+  int mode(unsigned int n) {
     return mode_[n];
   }
