+++ inst/include/snipsnop/SNPvectorMemory.h
@@ -55,7 +55,7 @@ class SNPvectorMemory : public SNPvector {
     return mode_;
   }
 
-  uint8_t mode(unsigned int n) {
+  int mode(unsigned int n) {
     return mode_[n];
   }
 };
